Refactor CyberGear and CyberGearV2: Remove error_code and result handling
- Deleted `error_code.hpp` and `result.hpp` files to simplify error handling.
- Updated `protocol_types.hpp` to remove dependencies on the deleted files.
- Refactored `cybergear.cpp` and `cybergear_v2.cpp` to eliminate Result<T, ErrorCode> usage, replacing it with direct error handling.
- Consolidated CAN frame building methods in both CyberGear and CyberGearV2 classes for better code organization.
- Enhanced data frame handling by updating the dispatch and update methods to improve clarity and maintainability.
- Adjusted the `data_frame_handler.cpp` to ensure proper frame type extraction without relying on removed error handling structures.
